Move basic file manipulation methods into own helpers file
